+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/intellij/codeInsight/hints/XPathInlayParameterHintsProvider.kt
@@ -23,8 +23,8 @@ import uk.co.reecedunn.intellij.plugin.xdm.functions.op.op_qname_presentation
 import uk.co.reecedunn.intellij.plugin.xdm.types.XdmAttributeNode
 import uk.co.reecedunn.intellij.plugin.xdm.types.XdmElementNode
 import uk.co.reecedunn.intellij.plugin.xdm.types.XsQNameValue
+import uk.co.reecedunn.intellij.plugin.xpath.ast.isArrowFunctionCall
 import uk.co.reecedunn.intellij.plugin.xpm.variable.XpmVariableName
-import uk.co.reecedunn.intellij.plugin.xpath.ast.plugin.PluginArrowFunctionCall
 import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.*
 import uk.co.reecedunn.intellij.plugin.xpm.context.expand
 
@@ -36,7 +36,7 @@ class XPathInlayParameterHintsProvider : InlayParameterHintsProvider {
             when {
                 binding.param.variableName == null -> null // Parameter with incomplete variable name.
                 binding.isEmpty() -> null // Empty variadic parameter.
-                index == 0 && element.parent is PluginArrowFunctionCall -> null // Arrow function call context argument.
+                index == 0 && element.parent.isArrowFunctionCall -> null // Arrow function call context argument.
                 getName(binding[0])?.localName?.data == binding.param.variableName?.localName?.data -> null
                 else -> op_qname_presentation(binding.param.variableName!!)?.let {
                     InlayInfo(it, binding[0].textOffset, false)