+++ src/lang-xpath/main/uk/co/reecedunn/intellij/plugin/xpath/intellij/lang/XPathParameterInfoHandler.kt
@@ -20,8 +20,8 @@ import com.intellij.codeInsight.lookup.LookupElement
 import com.intellij.lang.parameterInfo.*
 import com.intellij.psi.NavigatablePsiElement
 import uk.co.reecedunn.intellij.plugin.core.sequences.ancestors
+import uk.co.reecedunn.intellij.plugin.xpath.ast.isArrowFunctionCall
 import uk.co.reecedunn.intellij.plugin.xpm.function.XpmFunctionDeclaration
-import uk.co.reecedunn.intellij.plugin.xpath.ast.plugin.PluginArrowFunctionCall
 import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.*
 import uk.co.reecedunn.intellij.plugin.xpath.lexer.XPathTokenType
 import uk.co.reecedunn.intellij.plugin.xpath.model.staticallyKnownFunctions
@@ -36,7 +36,7 @@ class XPathParameterInfoHandler : ParameterInfoHandler<XPathArgumentList, XpmFun
         val args = e?.ancestors()?.filterIsInstance<XPathArgumentList>()?.firstOrNull()
         context.itemsToShow = functionCandidates(args).filter {
             if (it.arity.from == 0 && it.arity.to == 0)
-                args?.parent !is PluginArrowFunctionCall
+                args?.parent?.isArrowFunctionCall == false
             else
                 true
         }.toList().toTypedArray()
@@ -55,7 +55,7 @@ class XPathParameterInfoHandler : ParameterInfoHandler<XPathArgumentList, XpmFun
     override fun updateParameterInfo(parameterOwner: XPathArgumentList, context: UpdateParameterInfoContext) {
         val index =
             ParameterInfoUtils.getCurrentParameterIndex(parameterOwner.node, context.offset, XPathTokenType.COMMA)
-        context.setCurrentParameter(if (parameterOwner.parent is PluginArrowFunctionCall) index + 1 else index)
+        context.setCurrentParameter(if (parameterOwner.parent.isArrowFunctionCall) index + 1 else index)
     }
 
     override fun updateUI(p: XpmFunctionDeclaration?, context: ParameterInfoUIContext) {