+++ src/lang-xquery/test/uk/co/reecedunn/intellij/plugin/xquery/tests/intellij/codeInsight/hints/XQueryInlayParameterHintsProviderTest.kt
@@ -138,8 +138,13 @@ private class XQueryInlayParameterHintsProviderTest : ParserTestCase() {
             fun parenthesizedExpr() {
                 val args = parse<XPathArgumentList>("2 => (fn:abs#1)()")[0]
 
-                val info = provider.getHintInfo(args)
-                assertThat(info, `is`(nullValue()))
+                val info = provider.getHintInfo(args)!!
+                assertThat(info.fullyQualifiedName, `is`("Q{http://www.w3.org/2005/xpath-functions}abs"))
+                assertThat(info.language, `is`(nullValue()))
+                assertThat(info.getMethodName(), `is`("abs"))
+
+                assertThat(info.paramNames.size, `is`(1))
+                assertThat(info.paramNames[0], `is`("arg"))
             }
         }