+++ src/lang-xquery/test/uk/co/reecedunn/intellij/plugin/xquery/tests/intellij/lang/XQueryParameterInfoHandlerTest.kt
@@ -318,7 +318,10 @@ private class XQueryParameterInfoHandlerTest : ParserTestCase() {
                 assertThat(context.parameterListStart, `is`(16))
 
                 val items = context.itemsToShow!!.map { it as XpmFunctionDeclaration }
-                assertThat(items.size, `is`(0))
+                assertThat(items.size, `is`(1))
+
+                assertThat(op_qname_presentation(items[0].functionName!!), `is`("fn:abs"))
+                assertThat(items[0].arity, `is`(Range(1, 1)))
 
                 val hint = context as MockCreateParameterInfoContext
                 assertThat(hint.showHintElement, `is`(nullValue()))