+++ CHANGELOG.md
@@ -18,10 +18,14 @@ IntelliJ Integration:
     `DirElemContent`/`DirTextConstructor`, `CDataSection`, and `Pragma`
     elements.
 1.  Support `language=Language` injection marker comments.
-1.  Don't add inlay parameter hints when element and attribute node names
-    match the parameter name.
 1.  Use the tag tree highlight colours for XML names in the breadcrumb bar
     and editor.
+
+Inlay Parameters and Parameter Hints:
+
+1.  Don't add inlay parameter hints when element and attribute node names
+    match the parameter name.
+1.  Support expressions like `f#1(2)`, `(f#1(2))`, and `1 => (f#2(2))`.
 
 Saxon: